requirements.txt
pytest>=7.0
pandas>=1.3
pymysql>=1.0
python-dotenv>=0.20
black>=24.0
isort>=5.0
flake8>=6.0
matplotlib>=3.0
